pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-organizer"
version = "2.0.0"
authors = [
    {name = "Organisateur Photos Team"},
]
description = "Modern application to automatically sort your photos by date"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
]
keywords = ["photos", "organizer", "exif", "date", "sorting", "gui", "tkinter"]
dependencies = [
    "Pillow>=9.0.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/VOTRE_USERNAME/PhotoTransfer"
"Bug Reports" = "https://github.com/VOTRE_USERNAME/PhotoTransfer/issues"
"Source" = "https://github.com/VOTRE_USERNAME/PhotoTransfer"
"Documentation" = "https://github.com/VOTRE_USERNAME/PhotoTransfer/blob/main/README.md"

[project.scripts]
photo-organizer = "photo_organizer_v2:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["photo_organizer*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.json"]

# Black configuration (code formatting)
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# PyInstaller configuration via Python
[tool.pyinstaller]
name = "PhotoTransfer"
onefile = true
windowed = true
clean = true
noconfirm = true